move DedicatedResource/EvictionStrategy to src/utils/components
