+++ src/views/virtualmachines/details/tabs/details/components/grid/rightGrid/VirtualMachineSchedulingRightGrid.tsx
@@ -3,15 +3,15 @@ import * as React from 'react';
 import VirtualMachineModel from '@kubevirt-ui/kubevirt-api/console/models/VirtualMachineModel';
 import { V1VirtualMachine } from '@kubevirt-ui/kubevirt-api/kubevirt';
 import CPUMemoryModal from '@kubevirt-utils/components/CPUMemoryModal/CpuMemoryModal';
+import DedicatedResourcesModal from '@kubevirt-utils/components/DedicatedResourcesModal/DedicatedResourcesModal';
+import EvictionStrategyModal from '@kubevirt-utils/components/EvictionStrategyModal/EvictionStrategyModal';
 import { useModal } from '@kubevirt-utils/components/ModalProvider/ModalProvider';
 import { useKubevirtTranslation } from '@kubevirt-utils/hooks/useKubevirtTranslation';
 import { k8sUpdate } from '@openshift-console/dynamic-plugin-sdk';
 import { DescriptionList, GridItem } from '@patternfly/react-core';
 
 import DedicatedResources from '../../DedicatedResources/DedicatedResources';
-import DedicatedResourcesModal from '../../DedicatedResources/DedicatedResourcesModal';
 import EvictionStrategy from '../../EvictionStrategy/EvictionStrategy';
-import EvictionStrategyModal from '../../EvictionStrategy/EvictionStrategyModal';
 import Flavor from '../../Flavor/Flavor';
 import VirtualMachineDescriptionItem from '../../VirtualMachineDescriptionItem/VirtualMachineDescriptionItem';
